Course Project post (#87)
Added the missing information about the course Project.
Fixed the table bug.

I know that it's late, but I just wanted to explain that I believed
until now that I had fixed the bug, I did not realize it had failed
again. But I wanted to turn it in fixed and correct because I did not
want to see the error message or bug message.

I'm sorry for my confusion and I will be more aware next time.

---------

Co-authored-by: Gus Hahn-Powell <hahnpowell@arizona.edu>